[REPO-5317] * minor conventions fixes
