+++ packaging/tests/tas-restapi/src/test/java/org/alfresco/rest/models/aspects/GetAspectTests.java
@@ -2,152 +2,53 @@ package org.alfresco.rest.models.aspects;
 
 import org.alfresco.rest.RestTest;
 import org.alfresco.rest.model.RestAbstractClassModel;
-import org.alfresco.rest.model.RestAbstractClassModelsCollection;
+import org.alfresco.rest.model.RestErrorModel;
 import org.alfresco.utility.model.TestGroup;
-import org.alfresco.utility.model.UserModel;
 import org.alfresco.utility.testrail.ExecutionType;
 import org.alfresco.utility.testrail.annotation.TestRail;
 import org.springframework.http.HttpStatus;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-public class GetAspectTests  extends RestTest
+public class GetAspectTests extends RestTest
 {
-
-    private UserModel regularUser;
 
     @BeforeClass(alwaysRun=true)
     public void dataPreparation() throws Exception
     {
-        regularUser = dataUser.createRandomTestUser();
+        restClient.authenticateUser(dataUser.createRandomTestUser());
     }
 
     @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Verify user get aspects and gets status code OK (200)")
-    public void getAspects() throws Exception
+    @TestRail(section = { TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
+            description = "Verify inexistent aspect and status code is Not Found (404)")
+    public void getInexistentAspect() throws Exception
     {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.assertThat()
-            .entriesListCountIs(100)
-            .and().entriesListContains("id", "cm:classifiable")
-            .and().entriesListContains("id", "cm:author")
-            .and().entriesListContains("id", "cm:checkedOut");
+        String unknownAspect = "unknown:aspect";
+        restClient.withModelAPI().getAspect(unknownAspect);
+        restClient.assertStatusCodeIs(HttpStatus.NOT_FOUND)
+                .assertLastError().containsSummary(String.format(RestErrorModel.ENTITY_WAS_NOT_FOUND, unknownAspect));
     }
 
     @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Should filter aspects using namespace uri and gets status code OK (200)")
-    public void getAspectByNamespaceUri() throws Exception
+    @TestRail(section = { TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
+            description = "Verify Aspect Info and status code is OK (200)")
+    public void getAspect() throws Exception
     {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("where=(namespaceUri matches('http://www.alfresco.org/model.*'))")
-                .getAspects();
+        RestAbstractClassModel aspect = restClient.withModelAPI().getAspect("cm:titled");
         restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.assertThat().entriesListCountIs(100);
-
-        aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("where=(not namespaceUri matches('http://www.alfresco.org/model.*'))")
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.assertThat().entriesListCountIs(0);
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Should filter aspects using modelId and gets status code OK (200)")
-    public void getAspectByModelsIds() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("where=(modelId in ('cm:contentmodel', 'smf:smartFolder'))")
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.getPagination().assertThat().fieldsCount().is(5).and()
-                .field("totalItems").isLessThan(65).and()
-                .field("maxItems").is(100).and()
-                .field("skipCount").isGreaterThan(0).and()
-                .field("hasMoreItems").is(false);
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Should filter aspects using modelId with subaspects and gets status code OK (200)")
-    public void getAspectByModelsIdsWithIncludeSubAspects() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("where=(modelId in ('cm:contentmodel INCLUDESUBASPECTS', 'smf:smartFolder INCLUDESUBASPECTS'))")
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.getPagination().assertThat().fieldsCount().is(5).and()
-                .field("totalItems").isGreaterThan(65).and()
-                .field("maxItems").is(100).and()
-                .field("skipCount").isGreaterThan(0).and()
-                .field("hasMoreItems").is(false);
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Should filter aspects using parentId and gets status code OK (200)")
-    public void getAspectByParentId() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("where=(parentId in ('cm:titled'))")
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-        aspects.getPagination().assertThat().fieldsCount().is(5).and()
-                .field("totalItems").is(5).and()
-                .field("hasMoreItems").is(false);
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section = {TestGroup.REST_API, TestGroup.MODEL }, executionType = ExecutionType.REGRESSION,
-            description = "Should Aspects association, properties and mandatory aspects and gets status code OK (200)")
-    public void getAspectIncludeParams() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("include=properties,mandatoryAspects,associations")
-                .getAspects();
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-
-        for (RestAbstractClassModel aspect : aspects.getEntries())
-        {
-            aspect.onModel().assertThat()
-                    .field("associations").isNotNull().and()
-                    .field("properties").isNotNull().and()
-                    .field("mandatoryAspects").isNotNull();
-        }
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section={TestGroup.REST_API, TestGroup.MODEL}, executionType= ExecutionType.REGRESSION,
-            description= "Verify if any user gets aspects with high skipCount and maxItems parameter applied")
-    public void getPaginationParameter() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser)
-                .withModelAPI()
-                .usingParams("maxItems=10&skipCount=10")
-                .getAspects();
-        aspects.assertThat().entriesListCountIs(10);
-        aspects.assertThat().paginationField("hasMoreItems").is("true");
-        aspects.assertThat().paginationField("skipCount").is("10");
-        aspects.assertThat().paginationField("maxItems").is("10");
-        restClient.assertStatusCodeIs(HttpStatus.OK);
-    }
-
-    @Test(groups = { TestGroup.REST_API, TestGroup.MODEL, TestGroup.REGRESSION })
-    @TestRail(section={TestGroup.REST_API, TestGroup.MODEL}, executionType= ExecutionType.REGRESSION,
-            description= "Verify if any user gets aspects with hasMoreItems applied bases on skip count and maxItems")
-    public void getHighPaginationQuery() throws Exception
-    {
-        RestAbstractClassModelsCollection aspects = restClient.authenticateUser(regularUser).withModelAPI()
-                .usingParams("maxItems=10&skipCount=150")
-                .getAspects();
-        aspects.assertThat().entriesListCountIs(0);
-        aspects.assertThat().paginationField("hasMoreItems").is("false");
-        aspects.assertThat().paginationField("skipCount").is("150");
-        aspects.assertThat().paginationField("maxItems").is("10");
-        restClient.assertStatusCodeIs(HttpStatus.OK);
+        aspect.assertThat().field("associations").isEmpty().and()
+                .field("mandatoryAspects").isEmpty().and()
+                .field("properties").isNotEmpty().and()
+                .field("includedInSupertypeQuery").is(true).and()
+                .field("isContainer").is(false).and()
+                .field("id").is("cm:titled").and()
+                .field("description").is("Titled").and()
+                .field("title").is("Titled").and()
+                .field("model.id").is("cm:contentmodel").and()
+                .field("model.author").is("Alfresco").and()
+                .field("model.description").is("Alfresco Content Domain Model").and()
+                .field("model.namespaceUri").is("http://www.alfresco.org/model/content/1.0").and()
+                .field("model.namespacePrefix").is("cm");
     }
 }
